features/composer/index.js → ComposeForm
